requirements.txt
ccxt>=4.0.0
websockets>=12.0
requests>=2.31.0
python-dotenv>=1.0.0
flask>=3.0.0
flask-cors>=4.0.0

# Auto-posting dependencies
matplotlib>=3.7.0
pillow>=10.0.0
selenium>=4.15.0
tweepy>=4.14.0
